# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ImpactBridge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProblemStatus = (typeof ProblemStatus)[keyof typeof ProblemStatus];

export const ProblemStatus = {
  reported: "reported",
  community_joined: "community_joined",
  funding_started: "funding_started",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type ProblemUrgency =
  (typeof ProblemUrgency)[keyof typeof ProblemUrgency];

export const ProblemUrgency = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Problem {
  id: number;
  title: string;
  description: string;
  category: string;
  location: string;
  status: ProblemStatus;
  imageUrl?: string;
  fundingGoal: number;
  fundingRaised: number;
  progressPercent: number;
  joinedCount: number;
  createdAt: string;
  postedBy: string;
  postedByAvatar?: string;
  urgency: ProblemUrgency;
  verifiedCount?: number;
}

export interface Update {
  id: number;
  problemId: number;
  content: string;
  author: string;
  authorAvatar?: string;
  createdAt: string;
  imageUrl?: string;
}

export interface Comment {
  id: number;
  problemId: number;
  content: string;
  author: string;
  authorAvatar?: string;
  createdAt: string;
  likes?: number;
}

export interface UserSummary {
  id: number;
  name: string;
  avatarUrl?: string;
}

export type ProblemDetail = Problem & {
  images?: string[];
  updates?: Update[];
  comments?: Comment[];
  joinedUsers?: UserSummary[];
};

export interface CreateProblemBody {
  title: string;
  description: string;
  category: string;
  location: string;
  imageUrl?: string;
  fundingGoal: number;
  urgency?: string;
}

export interface CreateUpdateBody {
  content: string;
  imageUrl?: string;
}

export interface CreateCommentBody {
  content: string;
}

export interface Contribution {
  id: number;
  problemId: number;
  problemTitle: string;
  amount: number;
  userId: number;
  createdAt: string;
  anonymous?: boolean;
}

export interface CreateContributionBody {
  problemId: number;
  amount: number;
  anonymous?: boolean;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  bio?: string;
  avatarUrl?: string;
  location?: string;
  reputationScore: number;
  problemsCreated: number;
  problemsSolved: number;
  totalContributed: number;
  joinedAt: string;
}

export interface Notification {
  id: number;
  message: string;
  type: string;
  read: boolean;
  createdAt: string;
  problemId?: number;
}

export interface DashboardStats {
  totalContributions: number;
  problemsJoined: number;
  problemsCreated: number;
  reputationScore: number;
  myProblems: Problem[];
  myContributions: Contribution[];
  notifications: Notification[];
}

export interface CommunityGroup {
  id: number;
  name: string;
  description: string;
  memberCount: number;
  problemCount: number;
  category: string;
  imageUrl?: string;
  active: boolean;
  isMember?: boolean;
}

export interface MembershipResult {
  joined: boolean;
  alreadyMember?: boolean;
  wasMember?: boolean;
}

export type CommunityTaskStatus =
  (typeof CommunityTaskStatus)[keyof typeof CommunityTaskStatus];

export const CommunityTaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface CommunityTask {
  id: number;
  title: string;
  description: string;
  assignedTo?: string;
  assignedToAvatar?: string;
  problemId?: number;
  problemTitle?: string;
  status: CommunityTaskStatus;
  dueDate?: string;
}

export interface CommunityMessage {
  id: number;
  communityId: number;
  userId: number;
  userName: string;
  userAvatar?: string;
  message: string;
  createdAt: string;
}

export interface OverviewStats {
  totalProblems: number;
  problemsSolved: number;
  totalFundingRaised: number;
  activeCommunities: number;
  volunteersJoined: number;
}

export type VoteBodyVote = (typeof VoteBodyVote)[keyof typeof VoteBodyVote];

export const VoteBodyVote = {
  verify: "verify",
  dispute: "dispute",
} as const;

export interface VoteBody {
  vote: VoteBodyVote;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ListProblemsParams = {
  search?: string;
  category?: string;
  location?: string;
  status?: string;
  sort?: ListProblemsSort;
};

export type ListProblemsSort =
  (typeof ListProblemsSort)[keyof typeof ListProblemsSort];

export const ListProblemsSort = {
  newest: "newest",
  most_funded: "most_funded",
  most_urgent: "most_urgent",
} as const;

export type PostCommunityGroupMessageBody = {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  message: string;
};