# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ImpactBridge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /problems/{id}/comments:
    get:
      operationId: getProblemComments
      tags: [comments]
      summary: Get comments for a problem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a problem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /community/groups/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listCommunityGroupMessages
      tags: [community]
      summary: List chat messages for a community (members only)
      responses:
        "200":
          description: List of messages (oldest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityMessage"
        "403":
          description: Caller is not a member of this community
        "404":
          description: Community not found
    post:
      operationId: postCommunityGroupMessage
      tags: [community]
      summary: Post a chat message to a community (members only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required:
                - message
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityMessage"
        "403":
          description: Caller is not a member of this community
        "404":
          description: Community not found
  /community/groups/{id}/join:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: joinCommunityGroup
      tags: [community]
      summary: Join a community
      responses:
        "200":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipResult"
        "201":
          description: Newly joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipResult"
        "404":
          description: Community not found
    delete:
      operationId: leaveCommunityGroup
      tags: [community]
      summary: Leave a community
      responses:
        "200":
          description: Left (or was not a member)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipResult"
        "404":
          description: Community not found
...
components:
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        problemId:
          type: integer
        content:
          type: string
        author:
          type: string
        authorAvatar:
          type: string
        createdAt:
          type: string
        likes:
          type: integer
      required:
        - id
        - problemId
        - content
        - author
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    MembershipResult:
      type: object
      properties:
        joined:
          type: boolean
        alreadyMember:
          type: boolean
        wasMember:
          type: boolean
      required:
        - joined
...
    CommunityMessage:
      type: object
      properties:
        id:
          type: integer
        communityId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - communityId
        - userId
        - userName
        - message
        - createdAt
...